requirements.txt
torch
torchvision
transformers
diffusers
accelerate
streamlit
pillow
numpy
scikit-learn
